requirements.txt
# Cloud Custodian and AWS dependencies
c7n==0.9.36
c7n-org==0.6.29
boto3==1.34.51
botocore==1.34.51

# Additional utilities
pyyaml==6.0.1
python-dateutil==2.8.2
